Let albert toggle when invoking
In some occasions, albert loses focus, but does not close itself. By
then, it's still visible and the "show" command does not work.
If this occurs, toggle behaves better, as albert will disappear on the
first execution and appear **with** focus on the second one. "show"
would do nothing and you would have grab your mouse to focus albert
again.

For the normal use case, "toggle" and "show" work the same.