Set the desired terminal via TERMINAL env
Previously i3-sensible-terminal used /usr/bin/x-terminal-emulator and I
set it on my system manually. But this is not possible on NixOS.

While i3-sensible-terminal uses the TERMINAL to determine the user
preference, too, it's desired to use this. The xprofile file allows us
to export the TERMINAL variable upon session start, so that it's also
set in i3.

Why do we use i3-sensible-terminal? And not konsole directly? If konsole
is for $somereason not installed, opening a terminal would fail. While
i3-sensible-terminal would just use the next and at least just spawn a
terminal.