[git/config] Do not fetch the branches doubly
For some reason, previously git didn't merge the config and required to
have the fetch target of the branches stated explicitly in the
gitconfig. For some reason this changed in recent time and therefore git
has doubly branch information for fetching, resulting in a "multiple
updates not allowed" during clone.

The information is not necessary anymore, so it can be removed safely.