zsh autosuggestions: Remove match previous cmd
Since the last module update, the autosuggestions were a bit weird.
Apparently, my variable ZSH_AUTOSUGGESTIONS_STRATEGY was actually
obliged. In commit e614421 of zsh-autosuggestions, the config was
changed to actually oblige to the variable set in the zshrc.

So before the module update, we always used the standard "history"
strategy. During initial testing, I did not figure this out.

Since the module update, we really used match_prev_cmd as the
strategy (which actually sucks, I'm used to "history" now).

Additionally, I've activated "completion", which gives automatic
completion, if there is entry matching in the history.