git: replace diff-so-fancy with delta
diff-so-fancy does not work on `git add -p`, but delta does.

There had been a long discussion on the --patch issue in
diff-so-fancy[0] and the problems there, happen during git add, too.

Some commits [1] are referencing this issue. Here it got replaced with
delta.

After some checks, it seems to work fine and much better than
diff-so-fancy.

[0] https://github.com/so-fancy/diff-so-fancy/issues/35
[1] https://github.com/arcticicestudio/igloo/commit/48154199e6c08b954a49ed5df392b38c85a781b3